moving the audio logic into a dedicated module
