+++ noisekit/audio/__init__.py
@@ -1,0 +1,10 @@
+import threading
+from ..logging import get_logger
+
+
+class BaseThread(threading.Thread):
+
+    def __init__(self):
+        super().__init__()
+        self.shutdown_flag = threading.Event()
+        self.logger = get_logger(f"thread.{self.__class__.__name__}")
